+++ packages/contracts/src/Exception/EvolveException.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Evolve\Contracts\Exception;
+
+use Throwable;
+
+interface EvolveException extends Throwable {}