+++ packages/contracts/src/Lifecycle/ApplicationLifecycle.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Evolve\Contracts\Lifecycle;
+
+interface ApplicationLifecycle
+{
+    public function boot(): void;
+
+    public function shutdown(): void;
+}